+++ app/src/main/java/com/amazon/tv/leanbacklauncher/settings/LegacyAppOrderPreferenceFragment.java
@@ -1,13 +1,13 @@
 package com.amazon.tv.leanbacklauncher.settings;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v17.leanback.app.GuidedStepFragment;
-import android.support.v17.leanback.app.GuidedStepSupportFragment;
-import android.support.v17.leanback.widget.GuidanceStylist.Guidance;
-import android.support.v17.leanback.widget.GuidedAction;
-import android.support.v17.leanback.widget.GuidedAction.Builder;
-import android.support.v4.content.res.ResourcesCompat;
+
+import androidx.annotation.NonNull;
+import androidx.core.content.res.ResourcesCompat;
+import androidx.leanback.app.GuidedStepSupportFragment;
+import androidx.leanback.widget.GuidanceStylist.Guidance;
+import androidx.leanback.widget.GuidedAction;
+import androidx.leanback.widget.GuidedAction.Builder;
 
 import com.amazon.tv.leanbacklauncher.R;
 import com.amazon.tv.leanbacklauncher.apps.AppsManager;
@@ -24,12 +24,12 @@ public class LegacyAppOrderPreferenceFragment extends GuidedStepSupportFragment 
     public void onCreateActions(@NonNull List<GuidedAction> actions, Bundle savedInstanceState) {
         boolean z = true;
         SortingMode sortingMode = AppsManager.getSavedSortingMode(getActivity());
-        actions.add(new Builder(getActivity()).id(1).checkSetId(1).checked(sortingMode == SortingMode.RECENCY).title((int) R.string.select_app_order_action_description_recency).description((int) R.string.recency_order_action_description).build());
+        actions.add(new Builder(getActivity()).id(1).checkSetId(1).checked(sortingMode == SortingMode.RECENCY).title(R.string.select_app_order_action_description_recency).description(R.string.recency_order_action_description).build());
         Builder builder = new Builder(getActivity()).id(2).checkSetId(1);
         if (sortingMode != SortingMode.FIXED) {
             z = false;
         }
-        actions.add(builder.checked(z).title(R.string.select_app_order_action_description_fixed).description((int) R.string.fixed_order_action_description).build());
+        actions.add(builder.checked(z).title(R.string.select_app_order_action_description_fixed).description(R.string.fixed_order_action_description).build());
     }
 
     public void onGuidedActionClicked(GuidedAction action) {
